+++ Shared/Models/RefreshToken.cs
@@ -8,9 +8,9 @@ namespace Shared.Models
 {
     public class RefreshToken
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Token { get; set; } = string.Empty; // Token hasheado
-        public int EmployeeId { get; set; }
+        public long EmployeeId { get; set; }
         public DateTime Created { get; set; } = DateTime.UtcNow;
         public DateTime Expires { get; set; }
         public DateTime? Revoked { get; set; }
